requirements.txt
Flask==3.1.1
gunicorn==23.0.0
PyMySQL==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
email_validator==2.2.0
Flask-WTF==1.2.2
Flask-Admin==1.6.1
python-dotenv==1.1.1